garbage_api/requirements.txt
fastapi
uvicorn
pillow
numpy
tensorflow>=2.15
python-multipart
